mysite/requirements.txt
asgiref==3.3.4
Django==3.2.4
django-multiselectfield==0.1.12
django-simple-history==3.0.0
pytz==2021.1
sqlparse==0.4.1
